server/requirements.txt
Django==4.1.7
asgiref==3.6.0
gunicorn==20.1.0
sqlparse==0.4.3
tzdata==2022.7
psycopg2==2.9.5
stripe==5.3.0
django-cors-headers==3.14.0
django-datatables-view==1.20.0
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.0
websocket-client==1.6.3
daphne==4.0.0